+++ server/index.js
@@ -1,7 +1,7 @@
-import express from 'express'
-import cors from 'cors'
-import { spawn } from 'child_process'
-import {
+const express = require('express')
+const cors = require('cors')
+const { spawn } = require('child_process')
+const {
   S3Client,
   ListBucketsCommand,
   CreateBucketCommand,
@@ -11,8 +11,8 @@ import {
   PutObjectCommand,
   DeleteObjectCommand,
   DeleteObjectsCommand,
-} from '@aws-sdk/client-s3'
-import { getSignedUrl } from '@aws-sdk/s3-request-presigner'
+} = require('@aws-sdk/client-s3')
+const { getSignedUrl } = require('@aws-sdk/s3-request-presigner')
 
 const app = express()
 app.use(cors())